backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
docutils==0.21.1
filelock==3.13.4
Flask==3.0.3
h11==0.9.0
hpack==3.0.0
hyperframe==5.2.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
pillow==10.3.0
platformdirs==4.2.0
PyAudio==0.2.14
Pygments==2.17.2
pypiwin32==223
pywin32==306
requests==2.31.0
rfc3986==1.5.0
setuptools==69.5.1
soupsieve==2.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.2.1
vboxapi==1.0
virtualenv==20.25.1
Werkzeug==3.0.2
wheel==0.43.0
wikipedia==1.4.0
python-dotenv==1.0.0
